Fix doctrine/dbal 3 and doctrine/orm 2.10 compatibility (#6271)
* Replace json_array types with json types for doctrine/orm 2.10 support

* Fix PurgeDatabaseTrait for doctrine/dbal 3

* Fix TargetGroupCondition serialization

* Update phpstan baseline

* Update php-cs

* Fix setFirstResults and setMaxResults query builder methods

* Fix colleciton list pagination

* Fix collection view

+++ src/Sulu/Component/Rest/ListBuilder/Doctrine/DoctrineListBuilder.php
@@ -270,7 +270,7 @@ class DoctrineListBuilder extends AbstractListBuilder
 
         $subQueryBuilder = $this->createSubQueryBuilder($select);
         if (null != $this->limit) {
-            $subQueryBuilder->setMaxResults($this->limit)->setFirstResult($this->limit * ($this->page - 1));
+            $subQueryBuilder->setMaxResults((int) $this->limit)->setFirstResult((int) ($this->limit * ($this->page - 1)));
         }
 
         foreach ($this->sortFields as $index => $sortField) {
